+++ pipelex/cogt/ocr/ocr_job_factory.py
@@ -3,7 +3,6 @@ from typing import Optional
 from pipelex.cogt.ocr.ocr_input import OcrInput
 from pipelex.cogt.ocr.ocr_job import OcrJob
 from pipelex.cogt.ocr.ocr_job_components import OcrJobConfig, OcrJobParams, OcrJobReport
-from pipelex.config import get_config
 from pipelex.pipeline.job_metadata import JobCategory, JobMetadata
 
 
@@ -16,11 +15,9 @@ class OcrJobFactory:
         ocr_job_config: Optional[OcrJobConfig] = None,
         job_metadata: Optional[JobMetadata] = None,
     ) -> OcrJob:
-        config = get_config()
-        # TODO: manahge the param default sthrough the config
+        # TODO: manage the param default through the config
         # ocr_config = get_config().cogt.ocr_config
         job_metadata = job_metadata or JobMetadata(
-            top_job_id=f"OCRJob for {config.project_name}",
             job_category=JobCategory.OCR_JOB,
         )
         job_params = ocr_job_params or OcrJobParams.make_default_ocr_job_params()