+++ pipelex/pipe_controllers/pipe_parallel.py
@@ -102,4 +102,5 @@ class PipeParallel(PipeController):
                 )
         return PipeOutput(
             working_memory=working_memory,
+            pipeline_run_id=job_metadata.pipeline_run_id,
         )