+++ pipelex/pipe_operators/pipe_ocr.py
@@ -217,6 +217,7 @@ class PipeOcr(PipeOperator):
 
         pipe_output = PipeOcrOutput(
             working_memory=working_memory,
+            pipeline_run_id=job_metadata.pipeline_run_id,
         )
         return pipe_output